+++ CHANGES.md
@@ -6,6 +6,9 @@
   ([#118](https://github.com/kdgregory/log4j-aws-appenders/issues/118))
 * Can now use an `{ec2:tag:XXX}` substitution when running on EC2.
   ([#36](https://github.com/kdgregory/log4j-aws-appenders/issues/36))
+* CloudWatch and Kinesis log writers now expose the number of throttled
+  batches via JMX.
+  ([#127](https://github.com/kdgregory/log4j-aws-appenders/issues/127))
 * Can now use a `{uuid}` substitution.
   ([#128](https://github.com/kdgregory/log4j-aws-appenders/issues/128))
 * Now supports client endpoint configuration in all cases.